[Math] Fixes and refinements to Matrix/Vector generation

+++ Molten.Math/DoublePrecision/Matrices/Matrix4D.cs
@@ -302,7 +302,7 @@ namespace Molten.DoublePrecision
 			if (ptrValues == null)
 				throw new ArgumentNullException("ptrValues");
 
-			fixed (float* dst = Values)
+			fixed (double* dst = Values)
 				Unsafe.CopyBlock(ptrValues, dst, (sizeof(double) * 16));
 		}
